revamp hero
